08_board user 1:n model relation

+++ boards/views.py
@@ -17,7 +17,12 @@ def create(request):
         form = BoardForm(request.POST)
         #form 유효성 체크
         if form.is_valid():
-            board = form.save()
+            # board 를 바로 저장하지 않고, 현재 user를 넣고 저장
+            # 실제 DB에 반영 전까지의 단계를 진행하고, 그 중간에 user 정보를
+            # reqeust.user에서 가져와서 그 후에 저장한다.
+            board = form.save(commit=False) 
+            board.user = request.user
+            board.save()
             # 모델 폼은 검증이 필요 없어요. 이미 모델에서 검증.
             # title = form.cleaned_data.get('title')
             # content = form.cleaned_data.get('content')
@@ -38,31 +43,37 @@ def detail(request, board_pk):
     
 def delete(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        board.delete()
-        return redirect('boards:index')
+    if board.user == request.user:
+        if request.method == 'POST':
+            board.delete()
+            return redirect('boards:index')
+        else:
+            return redirect('boards:detail', board.pk)
     else:
-        return redirect('boards:detail', board.pk)
+        return redirect('boards:index')
 
 @login_required        
 def update(request, board_pk):
     board = get_object_or_404(Board, pk=board_pk)
-    if request.method == 'POST':
-        form = BoardForm(request.POST, instance=board)
-        if form.is_valid():
-            board = form.save()
-            # board.title = form.cleaned_data.get('title')
-            # board.content = form.cleaned_data.get('content')
-            # board.save()
-            return redirect('boards:detail', board.pk)
-    #GET 요청이면(수정하기 버튼을 눌렀을 때)
-    else:
-        #BoardForm을 초기화(사용자 입력 값을 넣어준 상태로)
-        #form = BoardForm(initial={'title':board.title, 'content':board.content})
-        form = BoardForm(instance=board)
-        # form = BoardForm(initial=board.__dict__)
-    #1. POST:요청에서 검증에 실패하였을 때, 오류 메세지가 포함된 상태
-    #2. GET : 요청에서 초기화된 상태
+    if board.user == request.user:
+        if request.method == 'POST':
+            form = BoardForm(request.POST, instance=board)
+            if form.is_valid():
+                board = form.save()
+                # board.title = form.cleaned_data.get('title')
+                # board.content = form.cleaned_data.get('content')
+                # board.save()
+                return redirect('boards:detail', board.pk)
+        #GET 요청이면(수정하기 버튼을 눌렀을 때)
+        else:
+            #BoardForm을 초기화(사용자 입력 값을 넣어준 상태로)
+            #form = BoardForm(initial={'title':board.title, 'content':board.content})
+            form = BoardForm(instance=board)
+            # form = BoardForm(initial=board.__dict__)
+        #1. POST:요청에서 검증에 실패하였을 때, 오류 메세지가 포함된 상태
+        #2. GET : 요청에서 초기화된 상태
+    else: 
+        return redirect('boards:index')
     context={
         'form':form, 
         'board':board,
